+++ simulator/js/state.js
@@ -40,7 +40,7 @@ const pendantMachine = {
   rotation: 2,
   spindleMaxRPM: 24000,
   spindleMinRPM: 0,
-  fluidDialVersion: "v2.0.2",
+  fluidDialVersion: "v2.0.3",
   fluidNCVersion: "v3.7.16",
   baudRate: "1000000",
   port: "UART0",